+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Upcoming v1.1.1
 - **FIX** some Tasmota 5.10.0 json #131 [hassio-addon#6](https://github.com/hassio-addons/addon-tasmoadmin/issues/6)
 - **UPDATE** Font Awesome 5.0.6 -> 5.1.0
+- **DOCKER** Show Changelog Link in helpmenu #108
 
 ## v1.1.0                              
 - **NEW**: you can now set a default startpage #115